add multiplayer functionalities
The game can be started at the same time on numerous clients, and they will be served the same first question.
After start game is pressed, the game sleeps for 3 seconds (introduce timer here), and then it begins simultaneously for both players.
Some distribution of logic was necessary for this step:
1) Numerous API calls at the same time would yield in multiple questions being created on the server, so in order to fix this I have introduced a "host", which is the player that starts the game
2) The host, requests for a new question, in advance (in order for the repo to process and have time to save it, for the next clients to come and get it)
3) The non-host players request for the question, only at UI change