add websocket functionality
The client-server communication now uses websockets instead of rest for the question transfer, lobby updates and start messages that benefit from a bidirectional protocol.

Added a new interface QuestionController that is implemented in the Question Scene Controllers so a reference to the current question controller is kept in client data.

Added a new class ResponseMessages that is used when sending data through websockets. 
Added a new constant ResponseCodes so after receiving a message through websockets the client will know what to do.

The client now doesn't refresh constantly, it receives data from the server whenever a player enters the lobby, the game starts or the question changes.

Some things implemented here are redundant and will be removed in future commits.

Some features are still not implemented such as updates when a player disconnects from the lobby.

Added an endpoint to add the player to lobby.