add local avatar generation and UI updates
The file generated gets named once, and then it only get accessed and his properties are changed according to the seed value. The seed value is computed using the players name, a base string (such that even for empty names, the generator would work), and a seed which gets incremented and decremented using the side buttons. This makes the user feel like he is truly changing his avatar based on his name. Moreover, the path to the user's avatar is saved in his class and stored on the repo so it can be used with other features such as leaderboards and emotes.
In order to compile the build with the new library, you might have to run CNTRL + SHIFT + O, on the build.gradle file. Right now, I skipped the renaming of the file part since I didn't find it being such a necessary feature (as long as the image itself is updated and stored correctly everything should be fine). I am also making use of temporary files which, from what I read, should automatically delete when the client exits. This doesn't happen right now, so I have a manual method "Clear avatars", that it is called once when the game is initialized which manually iterates through all the files in the "client/resources/images/avatar" directory and deletes them.