add the avatars to the waiting room as well
The process is easy to replicate for leaderboards and communication. The only thing that additionally is saved on the server is the string combination used to generate the seed for the avatar. The image generation is done client-sided (highly unlikely to put them on the repo since it would slow stuff done), making use of ExecutorService to schedule tasks: generation/creation of images from the provided seed. Please take a look at the code as I tried to document it as much as possible.