fix the file overload BUG and automatic renaming of files for better client-side storage management
There is one current BUG, where the avatar changes for one of the players in the lobby and is not matching the correct one, but I don't understand why that's happening right now. Before we get to the end of the project we will have to move everything that involves file management to regsitries, since we won't be able to use the same paths in our final build. I recommand this approach: https://docs.gradle.org/current/userguide/working_with_files.html, but we will have to dicuss this in a meeting