Show leaderboard at end of game and show avatar and rank
Now the avatar and rank are shown in every leaderboard.  However this done in a somewhat clunky way as I copied Simi's solution to show the avatars in the lobby, which is quite bulky to copy. I also added an attribute to LeaderboardEntry such that the avatar can be generated in all time high leaderboard. For this I removed avatarPath as it doesnt seem needed to me. Now all that needs to be done is persisting leaderboardentries after a singleplayer game.