fix the game breaking bug the made our back end a nightmare
The problem (questions being served twice in the same succession, together with a server-side error "Primary Key Violation", among multiple others) was solved with a bit of reformating. Instead of sending to the client a Question JSON that contains a list of activities, we send him a JSON that contains a list of (verified) activities IDS. In order to do so, I changed several structures both in the question class and the activity class (where now we have 2 ids: one generated automatically (a long), and the other fetched from the activities.json (String), which we don't use anymore). Then on the client side, when he receives the question, I make one more call to the server side to convert my activitiesIDs back into a list of activities that we use just like before.
This change had to be done because of the hibernate violation done on the primary keys when iniatiating any new type of collection of beans (which might point to an already existing id). In order to work-around it, I had to create a LongWrapper class, which is seen as a Bean entity that only holds a single value (the long id - from the activity id class).
All classes are used in the exact same way on the client side so we won't have to change any implementation aspects. Furthermore, we can use the generated lists of ids (on the client side), to send back to the server for every new request to make sure that no activities gets repeated.
Another thing I changed was the use of lists in the question class, to using sets (since these are more appropriate for database stuff).