Display text on GameMCQ through websocket
For now mostly in testing fase. Right now by pressing a button you can send a message that will pop up for every other player on the left side of the screen. The messages currently cycle through 3 spots, so there's a maximum of 3 messages displayed at most before the oldest one is overwritten.