Change message boxes to labels
Changed the messageTxt fields from Text to Label in order to display a background. Next I plan on trying to turn the message part into a mini scene such that it can be loaded into all the scenes where it is needed. If that doesn't work I'll copy what I have now to the other scenes where it is needed as a less elegant solution.