make the game end on the server side as well
I have also added some constants in the gameUtils class to keep track of when to serve the game over leaderboard and the temporary one. All endpoints and websockets work as discussed, and I have carefully tested them. The lobby is marked as inactive (notStarted), after the game ends, and a lobby gets ended automatically if there are no more players in it. A player cannot rejoin an on-going game.