change the lobby code on StartGame
Now the lobby code is changed with a random string when someone starts the game.
The client is communicated that the game has started and is provided with the new lobby token. After this, the client fetches the lobby through rest.

Fixed a bug where the variable currentSceneCtrl in clientutils was not updated from MainCtrl when a showEnergyAlternative scene appeared.