Fix bug regarding emotes
Fixed a bug that caused the emotes to not work anymore after loading a question type scene for the second time in one game. This caused the MenuItems to sometimes be added again to the MenuButton or something along those lines, as it was giving errors. By resetting the menubutton list of emotes everytime a question scene is loaded, this should not happen anymore.