make the half time joker work
The half time joker is now functional. The way it works is the following: 
1. Someone presses the half time joker and a message containing the joker type, the lobby token and the player's name is sent to /app/updateJoker
2. The message is then forwarded to /topic/updateJoker
3. All players subscribed receive the message and the ones that are in the lobby with the same token as the message check if the username in the message is equals to theirs and in the negative case the variable progress is halved. To keep all players in sync we don't check if progress <= 0 anymore but we have another variable timeLeft that we decrement in the timer. Once the progress bar depletes the answers become disabled so the player can no longer press them.

This commit also contains an example mechanic for all the jokers, it just seemed fun to play with it.

A bug where messages for jokers were not sent from the server to the client was fixed(see jokerController for the changes).