make it possible for the host of a private lobby to kick players out
the kicked player receives the message through a websocket and he is removed from the current lobby, the lobby is updated on the repo and his scene is set back to game mode selection screen. All the previous timers are killed (to not run in other bugs later on, even though my approach at it is a bit brutal (a null check in the refresh method in Waiting controller class ) ).
I have also introduced a new weboscket response code (KICK_PLAYER) appropriate for the situation